+++ camera_system/robot.py
